feat: `Url::path_is_root()` to determine if the path is `/`. (#450)
This could also be considered an empty path depending on the context
which is what makes it useful.
